ANALPA-2647: parameter change bug (#6)
* moved user input defaults to own file

* renamed function and added fieldValue prop for WayareaComponent

* added SelectedWayareaContext

* fixed bug where input field resets on initial render

* removed redundant prop

* added SelectedBoatContext

* fixed bug where input field resets after initial render
