+++ src/contexts/SelectedBoatContext.js
@@ -1,0 +1,5 @@
+const { createContext } = require("react");
+
+const SelectedBoatContext = createContext(null);
+
+export default SelectedBoatContext;
