+++ src/contexts/SelectedWayareaContext.js
@@ -1,0 +1,5 @@
+const { createContext } = require("react");
+
+const SelectedWayareaContext = createContext(null);
+
+export default SelectedWayareaContext;
